extract to external components + add sound and toast
