+++ app/routes/_layout-raw/_workshops/workshops_.$workshopSlug_.$slug/components/codante-logo-minimal.tsx
@@ -1,0 +1,36 @@
+export default function CodanteLogoMinimal() {
+  return (
+    <svg
+      width="39"
+      height="17"
+      viewBox="0 0 39 17"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M6.67289 0L9.67498 2.37007L4.8732 8.45233L9.67498 14.5346L6.67289 16.9047L0 8.45233L6.67289 0Z"
+        fill="#5282FF"
+      />
+      <path
+        className="dark:fill-white fill-gray-800"
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M10.4564 4.89722L14.3226 0L17.3247 2.37007L13.4585 7.26729C12.9099 7.9621 12.9099 8.94255 13.4585 9.63736L17.3247 14.5346L14.3226 16.9047L10.4564 12.0074C8.81077 9.92301 8.81077 6.98165 10.4564 4.89722Z"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M32.3271 0L29.325 2.37007L34.1268 8.45233L29.325 14.5346L32.3271 16.9047L39 8.45233L32.3271 0Z"
+        fill="#5282FF"
+      />
+      <path
+        className="dark:fill-white fill-gray-800"
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M28.5436 4.89722L24.6774 0L21.6753 2.37007L25.5415 7.26729C26.0901 7.9621 26.0901 8.94255 25.5415 9.63736L21.6753 14.5346L24.6774 16.9047L28.5436 12.0074C30.1892 9.92301 30.1892 6.98165 28.5436 4.89722Z"
+      />
+    </svg>
+  );
+}
